+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Bug fixes:
 - When recording conversation transcripts, Humbug was not recording the reasoning capabilities that were used in coming up
   with an answer.  This meant it was also not restoring these correctly after reloading a conversation.
 - The conversation input box could lose focus when an AI message finished streaming.
+- Markdown headings may now be preceeded by 0, 1, 2, or 3 spaces.
 
 ## v0.16 (2025-06-16)